+++ .gitignore
@@ -1,6 +1,6 @@
 node_modules
 
-# moccasin
+# sango
 dist/*
 !.gitkeep
