+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "moccasin",
+  "name": "sango",
   "version": "0.0.1",
   "description": "a leather shoe.",
   "scripts": {
