+++ projects/sango/index.spec.ts
@@ -3,7 +3,7 @@ import { hello } from "./"
 
 describe("hello", () => {
   it("should return hello", () => {
-    const message = hello("moccasin")
-    expect(message).to.eq("hello, moccasin!")
+    const message = hello("sango")
+    expect(message).to.eq("hello, sango!")
   })
 })
