+++ projects/sango/tsconfig.json
@@ -2,7 +2,7 @@
   "extends": "../tsconfig-base.json",
   "compilerOptions": {
     "baseUrl": ".",
-    "outDir": "../../dist/moccasin",
+    "outDir": "../../dist/sango",
     "composite": true
   }
 }
